Fix potential memory leaks after calling getline in OCSSD.
According to the man page of getline on Linux, "If *lineptr is set to
NULL and *n is set 0 before the call, then getline() will allocate a
buffer for storing the line.  This buffer should be freed by the user
program even if getline() failed."

When calling getline in OCSSD, we do intend to let getline allocate the
line buffers for us, but we forget to free them afterwards, hence
potential memory leaks.